# Speed up `import lumicks.pylake` by streamlining `scipy` imports