--- a/lumicks/pylake/force_calibration/tests/data/simulate_ideal.py
+++ b/lumicks/pylake/force_calibration/tests/data/simulate_ideal.py
@@ -1,6 +1,5 @@
 import numpy as np
-import scipy.signal
-import scipy.constants
+import scipy
 
 
 def calculate_active_term(time, driving_sinusoid, stiffness, gamma):