This follows one simple rule: At module scope, always use `import scipy`, never `imports scipy.<foo>` nor `from scipy import <foo>`.

Why: `scipy.<foo>` works anyway following `import scipy` and it will lazily import heavy submodules. Using `imports scipy.<foo>` forces the submodules to be imported eagerly.

Results: On my machine, cold import (first import following installation) goes down from 11.4 seconds to 3.9 seconds. Hot import goes down from 0.7 to 0.4 seconds.

Measuring: I followed the import time profiling method [described here](https://stackoverflow.com/a/51300944) plus a `pipenv` for reproducible cold import measurements:
```
mkdir tmp && cd tmp
pipenv --rm && pipenv shell
pip install <path/to/pylake>
python -X importtime -c "import lumicks.pylake" 2> pylake-cold.log
python -X importtime -c "import lumicks.pylake" 2> pylake-hot.log
tuna pylake-cold.log
tuna pylake-hot.log
```

#### Cold import: BEFORE

<img width="1112" alt="Screenshot 2023-08-08 at 10 02 59" src="https://github.com/lumicks/pylake/assets/4831417/15ff5d7d-7ebd-4350-8ad4-e00ffe10e999">

#### Cold import: AFTER

<img width="1115" alt="Screenshot 2023-08-08 at 10 03 14" src="https://github.com/lumicks/pylake/assets/4831417/3c264dca-cca4-4905-83d4-d9e82784209e">


#### Hot import: BEFORE

<img width="1115" alt="Screenshot 2023-08-08 at 10 03 56" src="https://github.com/lumicks/pylake/assets/4831417/15895a44-de53-4e5a-9305-6ef83ff027d5">

#### Hot import: AFTER

<img width="1115" alt="Screenshot 2023-08-08 at 10 04 08" src="https://github.com/lumicks/pylake/assets/4831417/1b0ca3c9-8796-4c96-b3e1-e095495d5b46">